refactor: acceptance 테스트 이름을 api 테스트로 변경
